requirements.txt
--index-url https://download.pytorch.org/whl/cu118

eel
pvporcupine
pyaudio
pyautogui
pywhatkit
wikipedia
requests
chromadb
psycopg
speedtest
google.generativeai
edge_tts
pygame
SpeechRecognition
torch
diffusers
playsound
yt_dlp
selenium
comtypes
pycaw
transformers
accelerate
torchvision
torchaudio
python-dotenv
tenacity